Make D2R usable with JNDI DataSources
This patch makes it possible to configure the database in the mapping
file with a JNDI DataSource. This is useful when deploying D2R as a war
file in Apache Tomcat, JBoss or some other servlet container.
Configuration of JDBC, connection pooling, etc is then shifted from
D2R to the container.