Address #164 by always applying --fast to DISTINCT
When --fast is off, we were conservatively assuming that DISTINCT is always
necessary. This is troublesome because LOBs don't support DISTINCT in
several RDBMS's. We now always apply the --fast logic to determine if
DISTINCT is necessary. As noted in the TripleRelationJoiner code, this
logic is incomplete and can fail.